mapper: put unroll code under crate::unroll
